Tackle memory issues. First the outputs are owned by the app but the other expressions are owned by the mixer. Inputs are always cloned so owned by the mixer. The mixer deletes its expressions so the output shouldnt delete its child expressions. In the funtab some strings were provided as literals as inbuilt functions. Since we want user defined functions, thses inbuilt functions have their string duplicated, else they have to be treated differently wen it comes to dleting them. I think it takes less memory to just duplicate them than to add stuff in the lookup_item to deal with the difference.
How c  strings are duplicated and deleted depends on the app so that is moved out of the library.
Similarly how error messages are printed depends on the app so only the declaration is provided. The app must fill in the definition.

Definitions of the above were moved out of the library then

The relevant function decls are moved to the lib header.

A make file is provided to create a stm32f405 version though currenetly depends on local paths